Adds test grouping to parallel test execution script

+++ parallel-test-runner.sh
@@ -4,49 +4,86 @@
 trap "kill 0" SIGINT SIGTERM EXIT
 
 function run_test_retry(){
-    local test=$1
+    local tests_file=$1
     local tmp_log_file=$2
     local i=0
     local exit_code=0
 
     while : ; do
         > $tmp_log_file
-        testr run --subunit $test > $tmp_log_file 2>&1
+        testr run --subunit --load-list=$tests_file > $tmp_log_file 2>&1
         subunit-stats $tmp_log_file > /dev/null
         exit_code=$?
         ((i++))
         ( [ $exit_code -eq 0 ] || [ $i -ge $retry_count ] ) && break
-        echo "Test $test failed. Retrying count: $i"
+        echo "Test $tests_file failed. Retrying count: $i"
     done
 
     echo $exit_code
 }
 
-function get_next_test_idx() {
+function get_tests_range() {
+    local i=$1
+    if [ $i -lt ${#tests[@]} ]; then
+        local test=${tests[$i]}
+        local test_class=${test%.*}
+        local j=$i
+        for test in ${tests[@]:$((i+1))}; do
+            local test_class_match=${test%.*}
+            if [ "$test_class" == "$test_class_match" ]; then
+                ((j++))
+            else
+                break
+            fi
+        done
+
+        echo $i $j
+    fi
+}
+
+function get_next_test_idx_range() {
    (
         flock -x 200
         local test_idx=$(<$cur_test_idx_file)
-        echo $test_idx
-        ((test_idx++))
-        echo $test_idx > $cur_test_idx_file
+        local test_idx_range=( $(get_tests_range $test_idx) )
+
+        if [ ${#test_idx_range[@]} -gt 0 ]; then
+            test_idx=${test_idx_range[1]}
+            ((test_idx++))
+            echo $test_idx > $cur_test_idx_file
+            echo ${test_idx_range[@]}
+        fi
    ) 200>$lock_file_1
 }
 
 function parallel_test_runner() {
     local runner_id=$1
-    while : ; do 
-        local test_idx=$(get_next_test_idx)
-        if [ $test_idx -ge ${#tests[@]} ]; then
+    while : ; do
+        local test_idx_range=( $(get_next_test_idx_range) )
+
+        if [ ${#test_idx_range[@]} -eq 0 ]; then
             break
         fi
-        local test=${tests[$test_idx]}
-        local tmp_log_file="$tmp_log_file_base"_"$test_idx"
 
-        echo "Test runner $runner_id is starting test $((test_idx+1)) out of ${#tests[@]}: $test"
+        local range_start=${test_idx_range[0]}
+        local range_end=${test_idx_range[1]}
+        local tmp_tests_file=$(tempfile)
+        local l=$((range_end-range_start+1))
 
-        local test_exit_code=$(run_test_retry $test $tmp_log_file)
-        
-        echo "Test runner $runner_id finished test $((test_idx+1)) out of ${#tests[@]} with exit code: $test_exit_code"
+        for test in ${tests[@]:$range_start:$l}; do
+            echo $test >> $tmp_tests_file
+        done
+
+        local tmp_log_file="$tmp_log_file_base"_"$range_start"
+
+        echo "Test runner $runner_id is starting tests from $((range_start+1)) to $((range_end+1)) out of ${#tests[@]}:"
+        cat $tmp_tests_file
+        echo
+
+        local test_exit_code=$(run_test_retry $tmp_tests_file $tmp_log_file)
+        rm $tmp_tests_file
+
+        echo "Test runner $runner_id finished test from $((range_start+1)) to $((range_end+1)) out of ${#tests[@]} with exit code: $test_exit_code"
     done
 }
 
@@ -80,8 +117,10 @@ rm $cur_test_idx_file
 > $log_file
 for i in $(seq 0 $((${#tests[@]}-1))); do
     tmp_log_file="$tmp_log_file_base"_"$i"
-    cat $tmp_log_file >> $log_file
-    rm $tmp_log_file
+    if [ -f "$tmp_log_file" ]; then
+        cat $tmp_log_file >> $log_file
+        rm $tmp_log_file
+    fi
 done
 
 rm $tmp_log_file_base
